fix error

+++ src/securityDevices/securityDevice.service.ts
@@ -94,9 +94,13 @@ export class SecurityDeviceService {
         if (userId === null) { return null }
         const deviceId = await  this.jwtService.getDeviceIdByRefreshToken(refreshToken)
         if (deviceId === null) { return null }
-
+        
         const issuedAt = await  this.jwtService.getIssuedAttByRefreshToken(refreshToken)
         if (issuedAt === null) { return null }
+
+        const validateIssuedAt = await this.securityDeviceRepository.findTokenAndDeviceByissuedAt(issuedAt)
+        if (validateIssuedAt === null) { return null }
+
         const resultDelete = await this.securityDeviceRepository.deleteTokenAndDevice(issuedAt)
         if (resultDelete === null) { return null }
